requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.0
opencv-python==4.7.0.72
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
urllib3==2.0.3
